+++ examples/postly/lib/domain/entity/post.dart
@@ -3,7 +3,7 @@ import 'package:freezed_annotation/freezed_annotation.dart';
 part 'post.freezed.dart';
 
 @freezed
-class Post with _$Post {
+abstract class Post with _$Post {
   const factory Post({
     required int id,
     required String title,