+++ examples/postly/lib/router/app_router.dart
@@ -5,8 +5,8 @@ import 'app_router.gr.dart';
 class AppRouter extends RootStackRouter {
   @override
   List<AutoRoute> get routes => [
-        AutoRoute(page: PostsRoute.page, initial: true),
-        AutoRoute(page: PostDetailsRoute.page),
-        AutoRoute(page: LogsRoute.page),
-      ];
+    AutoRoute(page: PostsRoute.page, initial: true),
+    AutoRoute(page: PostDetailsRoute.page),
+    AutoRoute(page: LogsRoute.page),
+  ];
 }